requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Deprecated==1.2.5
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.4
pep8==1.7.1
Pillow==6.0.0
powerline-status==2.7
PyGithub==1.43.7
PyJWT==1.7.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
webcolors==1.9.1
Werkzeug==0.15.4
wrapt==1.11.1
